Add test

+++ src/genos/fops.cpp
@@ -25,7 +25,7 @@ genos::openres * _get_schedee_file(int fd)
 	return (*sch->restbl)[fd];
 }
 
-ssize_t write(int fd, const void* data, size_t size) 
+int write(int fd, const void* data, size_t size) 
 {
 	genos::openres * filp;
 	genos::resource * res;
@@ -43,7 +43,7 @@ ssize_t write(int fd, const void* data, size_t size)
 	return node -> write(data, size, 0);
 }
 
-ssize_t read(int fd, void* data, size_t size) 
+int read(int fd, void* data, size_t size) 
 {
 	genos::openres * filp;
 	genos::resource * res;
